+++ lib/di.config.dart
@@ -12,18 +12,19 @@ import 'dart:async' as _i5;
 import 'dart:io' as _i6;
 
 import 'package:autogram_sign/autogram_sign.dart' as _i8;
-import 'package:eidmsdk/eidmsdk.dart' as _i14;
-import 'package:eidmsdk/types.dart' as _i13;
-import 'package:flutter/foundation.dart' as _i12;
+import 'package:eidmsdk/eidmsdk.dart' as _i15;
+import 'package:eidmsdk/types.dart' as _i14;
+import 'package:flutter/foundation.dart' as _i13;
 import 'package:get_it/get_it.dart' as _i1;
 import 'package:injectable/injectable.dart' as _i2;
 
 import 'app_service.dart' as _i3;
 import 'bloc/create_document_cubit.dart' as _i4;
-import 'bloc/present_signed_document_cubit.dart' as _i9;
-import 'bloc/preview_document_cubit.dart' as _i10;
-import 'bloc/select_signing_certificate_cubit.dart' as _i11;
-import 'bloc/sign_document_cubit.dart' as _i15;
+import 'bloc/paired_device_list_cubit.dart' as _i9;
+import 'bloc/present_signed_document_cubit.dart' as _i10;
+import 'bloc/preview_document_cubit.dart' as _i11;
+import 'bloc/select_signing_certificate_cubit.dart' as _i12;
+import 'bloc/sign_document_cubit.dart' as _i16;
 import 'data/pdf_signing_option.dart' as _i7;
 
 extension GetItInjectableX on _i1.GetIt {
@@ -48,39 +49,41 @@ extension GetItInjectableX on _i1.GetIt {
           file: file,
           pdfSigningOption: pdfSigningOption,
         ));
-    gh.factoryParam<_i9.PresentSignedDocumentCubit,
+    gh.factory<_i9.PairedDeviceListCubit>(
+        () => _i9.PairedDeviceListCubit(service: gh<_i8.IAutogramService>()));
+    gh.factoryParam<_i10.PresentSignedDocumentCubit,
         _i8.SignDocumentResponseBody, dynamic>((
       signedDocument,
       _,
     ) =>
-        _i9.PresentSignedDocumentCubit(
+        _i10.PresentSignedDocumentCubit(
           appService: gh<_i3.AppService>(),
           signedDocument: signedDocument,
         ));
-    gh.factoryParam<_i10.PreviewDocumentCubit, String, dynamic>((
+    gh.factoryParam<_i11.PreviewDocumentCubit, String, dynamic>((
       documentId,
       _,
     ) =>
-        _i10.PreviewDocumentCubit(
+        _i11.PreviewDocumentCubit(
           service: gh<_i8.IAutogramService>(),
           documentId: documentId,
         ));
-    gh.factoryParam<_i11.SelectSigningCertificateCubit,
-        _i12.ValueNotifier<_i13.Certificate?>, dynamic>((
+    gh.factoryParam<_i12.SelectSigningCertificateCubit,
+        _i13.ValueNotifier<_i14.Certificate?>, dynamic>((
       signingCertificate,
       _,
     ) =>
-        _i11.SelectSigningCertificateCubit(
-          eidmsdk: gh<_i14.Eidmsdk>(),
+        _i12.SelectSigningCertificateCubit(
+          eidmsdk: gh<_i15.Eidmsdk>(),
           signingCertificate: signingCertificate,
         ));
-    gh.factoryParam<_i15.SignDocumentCubit, String, _i13.Certificate>((
+    gh.factoryParam<_i16.SignDocumentCubit, String, _i14.Certificate>((
       documentId,
       certificate,
     ) =>
-        _i15.SignDocumentCubit(
+        _i16.SignDocumentCubit(
           service: gh<_i8.IAutogramService>(),
-          eidmsdk: gh<_i14.Eidmsdk>(),
+          eidmsdk: gh<_i15.Eidmsdk>(),
           documentId: documentId,
           certificate: certificate,
         ));